refactor part1
