+++ admin-frontend/app_singleapp/lib/widgets/features/table-collapsed-view/value_boolean_cell.dart
@@ -1,5 +1,5 @@
-import 'package:app_singleapp/widgets/features/dashboard-values-containers/value_not_set_container.dart';
 import 'package:app_singleapp/widgets/features/feature_value_status_tags.dart';
+import 'package:app_singleapp/widgets/features/table-collapsed-view/value_not_set_container.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:mrapi/api.dart';
