+++ admin-frontend/app_singleapp/lib/widgets/features/table-collapsed-view/value_not_set_container.dart
